Bug fixes, docs, migrate to model
